refactor of src modules
