+++ docs/Manifest.toml
@@ -2,7 +2,7 @@
 
 julia_version = "1.11.1"
 manifest_format = "2.0"
-project_hash = "ae700efc4a79ee7ce47a645a49f512199d5eddc2"
+project_hash = "be80539e480ad6014bf73104a780145f7797ea71"
 
 [[deps.ADTypes]]
 git-tree-sha1 = "72af59f5b8f09faee36b4ec48e014a79210f2f4f"
