+++ docs/Project.toml
@@ -1,5 +1,4 @@
 [deps]
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 BAT = "c0cd4b16-88b7-57fa-983b-ab80aecada7e"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
@@ -7,6 +6,7 @@ Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 FilePathsBase = "48062228-2e41-5def-b9a4-89aafe57970f"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
@@ -27,6 +27,7 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PropDicts = "4dc08600-4268-439e-8673-d706fafbb426"
 PropertyFunctions = "09e99361-2bb8-48a2-a80f-de58f0739eb4"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
