+++ docs/src/api.md
@@ -1,0 +1,118 @@
+# utils.jl Documentation
+
+```@docs
+ZeroNuFit.Utils.check_key
+ZeroNuFit.Utils.get_settings
+ZeroNuFit.Utils.get_events
+ZeroNuFit.Utils.event_is_contained
+ZeroNuFit.Utils.get_partitions
+ZeroNuFit.Utils.get_partitions_new
+ZeroNuFit.Utils.get_partitions_events
+ZeroNuFit.Utils.get_partition_event_index
+ZeroNuFit.Utils.get_corr_info
+ZeroNuFit.Utils.get_energy_scale_pars
+ZeroNuFit.Utils.get_efficiency
+ZeroNuFit.Utils.get_deltaE
+ZeroNuFit.Utils.get_range
+ZeroNuFit.Utils.get_bkg_info
+ZeroNuFit.Utils.save_generated_samples
+ZeroNuFit.Utils.save_results_into_json
+ZeroNuFit.Utils.get_global_mode
+ZeroNuFit.Utils.get_marginalized_mode
+
+ZeroNuFit.Plotting.plot_data
+ZeroNuFit.Plotting.plot_correlation_matrix
+ZeroNuFit.Plotting.plot_marginal_distr
+
+ZeroNuFit.Likelihood.get_stat_blocks
+ZeroNuFit.Likelihood.run_fit_over_partitions
+ZeroNuFit.Likelihood.get_signal_prior_info
+ZeroNuFit.Likelihood.norm_uniform
+ZeroNuFit.Likelihood.norm_linear
+ZeroNuFit.Likelihood.norm_exponential
+ZeroNuFit.Likelihood.gaussian_plus_lowEtail
+ZeroNuFit.Likelihood.get_mu_b
+ZeroNuFit.Likelihood.get_mu_s
+ZeroNuFit.Likelihood.get_mu_s_b
+ZeroNuFit.Likelihood.build_likelihood_zero_obs_evts
+ZeroNuFit.Likelihood.build_likelihood_per_partition
+ZeroNuFit.Likelihood.build_likelihood_looping_partitions
+ZeroNuFit.Likelihood.generate_data
+ZeroNuFit.Likelihood.get_signal_bkg_priors
+ZeroNuFit.Likelihood.build_prior
+
+ZeroNuFit.Likelihood.Utils.get_efficiency 
+ZeroNuFit.Analysis.Utils.get_efficiency 
+ZeroNuFit.Analysis.Likelihood.Utils.get_deltaE 
+ZeroNuFit.Analysis.Utils.get_partitions_new 
+ZeroNuFit.Analysis.Likelihood.Utils.get_marginalized_mode 
+ZeroNuFit.Likelihood.Utils.get_partitions_events 
+ZeroNuFit.Analysis.Likelihood.Utils.get_partitions_events 
+ZeroNuFit.Analysis.Likelihood.Utils.get_partition_event_index 
+ZeroNuFit.Analysis.Utils.get_bkg_info 
+ZeroNuFit.Likelihood.Utils.get_events 
+ZeroNuFit.Analysis.Likelihood.Utils.get_partitions_new 
+ZeroNuFit.Analysis.save_outputs 
+ZeroNuFit.Analysis.Likelihood.Utils.check_key 
+ZeroNuFit.Analysis.Likelihood.Utils.get_energy_scale_pars 
+ZeroNuFit.Analysis.Likelihood.run_fit_over_partitions 
+ZeroNuFit.Likelihood.Utils.get_partitions_new 
+ZeroNuFit.Analysis.Likelihood.Utils.get_efficiency 
+ZeroNuFit.Analysis.Likelihood.build_likelihood_zero_obs_evts 
+ZeroNuFit.Analysis.Likelihood.build_likelihood_looping_partitions
+ZeroNuFit.Analysis.Likelihood.gaussian_plus_lowEtail 
+ZeroNuFit.Analysis.Utils.get_partition_event_index 
+ZeroNuFit.Likelihood.Utils.get_partitions 
+ZeroNuFit.Analysis.Likelihood.norm_uniform 
+ZeroNuFit.Analysis.Likelihood.get_signal_prior_info 
+ZeroNuFit.Analysis.Utils.save_generated_samples 
+ZeroNuFit.Likelihood.Utils.get_partition_event_index 
+ZeroNuFit.Analysis.Plotting.plot_data 
+ZeroNuFit.Likelihood.Utils.get_global_mode 
+ZeroNuFit.Analysis.Utils.event_is_contained 
+ZeroNuFit.Analysis.Likelihood.Utils.get_range 
+ZeroNuFit.Analysis.Likelihood.Utils.get_bkg_info 
+ZeroNuFit.Likelihood.Utils.get_corr_info 
+ZeroNuFit.Likelihood.Utils.get_settings 
+ZeroNuFit.Analysis.Likelihood.Utils.save_results_into_json 
+ZeroNuFit.Analysis.Utils.get_partitions_events 
+ZeroNuFit.Analysis.Likelihood.Utils.event_is_contained 
+ZeroNuFit.Analysis.Likelihood.get_mu_b 
+ZeroNuFit.Analysis.Utils.get_events 
+ZeroNuFit.Analysis.Likelihood.Utils.get_global_mode 
+ZeroNuFit.Analysis.Likelihood.get_mu_s 
+ZeroNuFit.Likelihood.Utils.check_key 
+ZeroNuFit.Likelihood.Utils.get_bkg_info 
+ZeroNuFit.Analysis.Likelihood.norm_linear
+ZeroNuFit.Analysis.Likelihood.Utils.get_settings 
+ZeroNuFit.Analysis.Utils.save_results_into_json 
+ZeroNuFit.Analysis.Utils.get_settings 
+ZeroNuFit.Likelihood.Utils.save_generated_samples 
+ZeroNuFit.Analysis.Plotting.plot_marginal_distr 
+ZeroNuFit.Analysis.Utils.get_corr_info 
+ZeroNuFit.Analysis.Likelihood.Utils.get_partitions 
+ZeroNuFit.Analysis.Utils.get_deltaE 
+ZeroNuFit.Analysis.Likelihood.norm_exponential
+ZeroNuFit.Analysis.Utils.get_energy_scale_pars 
+ZeroNuFit.Likelihood.Utils.get_deltaE 
+ZeroNuFit.Likelihood.Utils.event_is_contained 
+ZeroNuFit.Analysis.Plotting.plot_correlation_matrix 
+ZeroNuFit.Analysis.Utils.get_range 
+ZeroNuFit.Analysis.Likelihood.build_likelihood_per_partition 
+ZeroNuFit.Likelihood.Utils.get_energy_scale_pars 
+ZeroNuFit.Analysis.Likelihood.Utils.get_events 
+ZeroNuFit.Likelihood.Utils.save_results_into_json 
+ZeroNuFit.Likelihood.Utils.get_range 
+ZeroNuFit.Analysis.Likelihood.build_prior 
+ZeroNuFit.Analysis.Utils.get_partitions 
+ZeroNuFit.Analysis.Utils.get_marginalized_mode 
+ZeroNuFit.Analysis.Utils.get_global_mode 
+ZeroNuFit.Analysis.Likelihood.generate_data 
+ZeroNuFit.Analysis.Utils.check_key 
+ZeroNuFit.Analysis.Likelihood.get_mu_s_b 
+ZeroNuFit.Likelihood.Utils.get_marginalized_mode 
+ZeroNuFit.Analysis.Likelihood.Utils.save_generated_samples 
+ZeroNuFit.Analysis.Likelihood.get_stat_blocks
+ZeroNuFit.Analysis.Likelihood.Utils.get_corr_info 
+ZeroNuFit.Analysis.Likelihood.get_signal_bkg_priors
+```
