+++ test/io/test_get_corr_info.jl
@@ -4,11 +4,10 @@ Pkg.instantiate()
 using Random
 include("../../src/ZeroNuFit.jl")
 using .ZeroNuFit
-include("../../main.jl")
 
 @testset "test_get_corr_info" begin
 
-    @info "Testing function to retrieve bkg correlation info (function 'get_corr_info' in src/fitting.jl)"
+    @info "Testing function to retrieve bkg correlation info (function 'get_corr_info' in src/utils.jl)"
 
     # no entry for correlated bkg
     config = Dict("bkg" => Dict())
@@ -17,7 +16,7 @@ include("../../main.jl")
     hier_mode = nothing
     hier_range = nothing
     try
-        corr, hier_mode, hier_range = ZeroNuFit.get_corr_info(config)
+        corr, hier_mode, hier_range = ZeroNuFit.Utils.get_corr_info(config)
     catch e
         @error "Error in 'get_corr_info' evaluation: $e"
         throw(e)
@@ -34,7 +33,7 @@ include("../../main.jl")
     hier_mode = nothing
     hier_range = nothing
     try
-        corr, hier_mode, hier_range = ZeroNuFit.get_corr_info(config)
+        corr, hier_mode, hier_range = ZeroNuFit.Utils.get_corr_info(config)
     catch e
         @error "Error in 'get_corr_info' evaluation: $e"
         throw(e)
@@ -53,7 +52,7 @@ include("../../main.jl")
     hier_mode = nothing
     hier_range = nothing
     try
-        corr, hier_mode, hier_range = ZeroNuFit.get_corr_info(config)
+        corr, hier_mode, hier_range = ZeroNuFit.Utils.get_corr_info(config)
     catch e
         @error "Error in 'get_corr_info' evaluation: $e"
         throw(e)
