+++ test/io/test_get_signal_prior_info.jl
@@ -4,11 +4,10 @@ Pkg.instantiate()
 using Random
 include("../../src/ZeroNuFit.jl")
 using .ZeroNuFit
-include("../../main.jl")
 
 @testset "test_get_signal_prior_info" begin
 
-    @info "Testing function to retrieve signal prior info (function 'get_signal_prior_info' in src/fitting.jl)"
+    @info "Testing function to retrieve signal prior info (function 'get_signal_prior_info' in src/likelihood.jl)"
 
     # flat S prior (default)
     config = Dict("signal" => Dict("prior" => "uniform", "upper_bound" => 10))
@@ -18,7 +17,7 @@ include("../../main.jl")
     sqrt_prior = nothing
     s_max = nothing
     try
-        sqrt_prior, s_max = ZeroNuFit.get_signal_prior_info(bkg_only, config)
+        sqrt_prior, s_max = ZeroNuFit.Likelihood.get_signal_prior_info(bkg_only, config)
     catch e
         @error "Error in 'get_signal_prior_info' evaluation: $e"
         throw(e)
@@ -32,7 +31,7 @@ include("../../main.jl")
     sqrt_prior = nothing
     s_max = nothing
     try
-        sqrt_prior, s_max = ZeroNuFit.get_signal_prior_info(bkg_only, config)
+        sqrt_prior, s_max = ZeroNuFit.Likelihood.get_signal_prior_info(bkg_only, config)
     catch e
         @error "Error in 'get_signal_prior_info' evaluation: $e"
         throw(e)
@@ -48,7 +47,7 @@ include("../../main.jl")
     sqrt_prior = nothing
     s_max = nothing
     try
-        sqrt_prior, s_max = ZeroNuFit.get_signal_prior_info(bkg_only, config)
+        sqrt_prior, s_max = ZeroNuFit.Likelihood.get_signal_prior_info(bkg_only, config)
     catch e
         @error "Error in 'get_signal_prior_info' evaluation: $e"
         throw(e)
