+++ test/statistics/test_inverse_uniform_cdf.jl
@@ -4,8 +4,6 @@ Pkg.instantiate()
 using Random
 include("../../src/ZeroNuFit.jl")
 using .ZeroNuFit
-include("../../main.jl")
-include("../../src/utils.jl")
 
 @testset "test_inverse_uniform_cdf" begin
 
@@ -17,7 +15,7 @@ include("../../src/utils.jl")
 
     res = nothing
     try
-        res = ZeroNuFit.inverse_uniform_cdf(n, fit_range)
+        res = ZeroNuFit.Utils.inverse_uniform_cdf(n, fit_range)
     catch e
         @error "Error in 'inverse_uniform_cdf' evaluation: $e"
         throw(e)
@@ -35,7 +33,7 @@ include("../../src/utils.jl")
     end
 
     # lower edge
-    res = ZeroNuFit.inverse_uniform_cdf(0, fit_range)
+    res = ZeroNuFit.Utils.inverse_uniform_cdf(0, fit_range)
     expected_value = 1
     tolerance = 1e-5
     @testset "Check res accuracy at lower edge" begin
@@ -43,7 +41,7 @@ include("../../src/utils.jl")
         @test diff <= tolerance
     end
     # upper edge
-    res = ZeroNuFit.inverse_uniform_cdf(1, fit_range)
+    res = ZeroNuFit.Utils.inverse_uniform_cdf(1, fit_range)
     expected_value = 30
     tolerance = 1e-5
     @testset "Check res accuracy at upper edge" begin
@@ -52,7 +50,7 @@ include("../../src/utils.jl")
     end
 
     # simple case
-    res = ZeroNuFit.inverse_uniform_cdf(0.5, [[0, 1]])
+    res = ZeroNuFit.Utils.inverse_uniform_cdf(0.5, [[0, 1]])
     expected_value = 0.5
     tolerance = 1e-5
     @testset "Check res accuracy in [0,1] at p=0.5" begin
